Added final version of 1.0
This version is the first public release. The App folder contains the Menu Bar app whilst the Helper folder contains the helper tool that runs when the App calls on it. The autolaunch folder contains the .plist that should be added to /Library/LaunchAgents to launch the app on login.